project details bar standalone comp
