+++ app/(dashboard)/[workspaceId]/[projectId]/layout.tsx
@@ -4,23 +4,16 @@ import { useState, useEffect } from 'react';
 // next
 import { useParams, useRouter } from 'next/navigation';
 // MUI
-import { Box, Typography, Tooltip, IconButton, Stack, Skeleton } from '@mui/material';
-import { Edit as EditIcon, CalendarMonth as CalendarMonthIcon } from '@mui/icons-material';
-// utils
-import { getLocalTimeZone, today, CalendarDate } from '@internationalized/date';
-import { getRelativeTime } from '@utils/datetime';
-// our components
+import { Box, Skeleton } from '@mui/material';
+// comps
 import GanttChart from '@components/GanttChart/GanttChart';
-import { useWorkspacesManager } from '@hooks/WorkspacesContext';
-import ProjectForm from '@components/Project/ProjectForm';
 import SprintDashboard from '@components/Sprint/Dashboard';
 import Drawer from '@UI/Drawer/Drawer';
+import ProjectDetailsBar from '@/app/components/Project/ProjectDetailsBar';
+// contexts
+import { useWorkspacesManager } from '@hooks/WorkspacesContext';
 // const
 import { project_details_bar_height, appbar_height, mock_elapse } from '@const';
-// styles
-import styles from './ProjectDashboard.module.css';
-
-const fallbackDesc = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.';
 
 /********************************************************************************************************************
  * project dashboard
@@ -32,14 +25,12 @@ export default function ProjectLayout({ children }: { children: React.ReactNode 
     sprintId?: string;
   };
   // contexts
-  const { workspaces, getProject, updateProjectFields } = useWorkspacesManager();
+  const { workspaces, getProject } = useWorkspacesManager();
   const workspace = workspaces[workspaceId];
   const project = getProject(workspaceId, projectId);
   const router = useRouter();
 
   // UI
-  const [editDialogOpen, setEditDialogOpen] = useState(false);
-  const [showFullDesc, setShowFullDesc] = useState(false);
   const [drawerOpen, setDrawerOpen] = useState(false);
   const [loading, setLoading] = useState(true);
 
@@ -63,89 +54,6 @@ export default function ProjectLayout({ children }: { children: React.ReactNode 
   if (!workspace || !project) {
     return <div>Invalid workspace or project</div>;
   }
-
-  /******************************************************************************************************************
-   * edit project
-   ******************************************************************************************************************/
-  const handleEditProject = async (title: string, desc: string, dueDate: CalendarDate) => {
-    if (workspaceId && projectId) {
-      await new Promise(res => setTimeout(res, mock_elapse));
-      updateProjectFields(workspaceId, projectId, { title, desc, dueDate });
-    }
-  };
-
-  /******************************************************************************************************************
-   * project details
-   ******************************************************************************************************************/
-  const projectDesc = (description: string) => (
-    <Box>
-      <Typography
-        variant='body2'
-        color='text.secondary'
-        className={`${styles.projectDesc} ${!showFullDesc ? styles.projectDescCollapsed : ''}`}
-      >
-        {description}
-      </Typography>
-
-      {description.length > 200 && (
-        <Typography
-          variant='caption'
-          color='primary'
-          sx={{ cursor: 'pointer', mt: 0.5, display: 'inline-block' }}
-          onClick={() => setShowFullDesc(prev => !prev)}
-        >
-          {showFullDesc ? 'Show less' : 'Show more'}
-        </Typography>
-      )}
-    </Box>
-  );
-
-  const projectDetailsBar = () => {
-    const dueDate: CalendarDate = project.dueDate;
-    const now: CalendarDate = today(getLocalTimeZone());
-
-    const formattedDue = `${dueDate.day}/${dueDate.month}/${dueDate.year}`;
-    const relativeDue = getRelativeTime(now, dueDate);
-    const description = project.desc || fallbackDesc;
-    // const description = fallbackDesc;
-
-    return (
-      <Box sx={{ display: 'flex', alignItems: 'flex-start', justifyContent: 'space-between', p: 2, minHeight: project_details_bar_height }}>
-        <Box sx={{ flex: 1 }}>
-          {!loading ?
-            <Box sx={{ height: 32 }}>
-              <Typography variant='h5' fontWeight={600}>
-                {`${workspace.title} - ${project.title}`}
-              </Typography>
-            </Box> :
-            <Skeleton width='50%' height={32} />}
-
-          <Stack sx={{ height: 22 }} direction='row' alignItems='center' spacing={0.5} mt={1}>
-            <CalendarMonthIcon fontSize='small' color='action' />
-            {!loading ?
-              <Typography variant='subtitle2' color='text.secondary'>
-                {`Complete by: ${formattedDue} (${relativeDue})`}
-              </Typography> :
-              <Skeleton width='30%' height={22} />}
-          </Stack>
-
-          {!loading ?
-            <Box sx={{ mt: 2, mr: 6 }}>
-              {projectDesc(description)}
-            </Box> :
-            <Box sx={{ mt: 1 }}>
-              <Skeleton width='100%' height={32} />
-            </Box>}
-        </Box>
-
-        <Tooltip title='Edit Project'>
-          <IconButton onClick={() => setEditDialogOpen(true)} sx={{ mt: 0.5 }}>
-            <EditIcon />
-          </IconButton>
-        </Tooltip>
-      </Box>
-    );
-  };
 
   /******************************************************************************************************************
    * sprint
@@ -180,16 +88,13 @@ export default function ProjectLayout({ children }: { children: React.ReactNode 
         />
       </Drawer>
 
-      {/* edit project form */}
-      {project ? <ProjectForm
-        workspace={workspaces[workspaceId]}
+      {/* project details bar */}
+      <ProjectDetailsBar
+        workspaceId={workspaceId}
+        workspaceTitle={workspaces[workspaceId].title}
         project={project}
-        projectDialogOpen={editDialogOpen}
-        onSubmitProject={handleEditProject}
-        closeProjectDialog={() => setEditDialogOpen(false)} /> : null}
-
-      {/* project details top bar */}
-      {projectDetailsBar()}
+        loading={loading}
+      />
 
       {/* project Gantt chart */}
       {!loading ?
