+++ app/components/Project/ProjectDetailsBar.module.css
